+++ www/go/core/Language.php
@@ -64,13 +64,23 @@ class Language {
 		return $this->isoCode;
 	}
 
+
+	public function getTextDirection(): string
+	{
+		if(in_array($this->getIsoCode(), ['ar', 'he', 'ur'])) {
+			return 'rtl';
+		}
+		return 'ltr';
+	}
+
+
 	/**
 	 * Set new language
 	 *
 	 * @param string|null $isoCode when not given it's detected from the browser
 	 * @return string|false Old language setting
 	 */
-	public function setLanguage($isoCode = null) {
+	public function setLanguage(?string $isoCode = null) {
 		$old = $this->getIsoCode();
 		if(!isset($isoCode)) {
 			$isoCode = $this->getBrowserLanguage();