+++ www/go/base/mail/Imap.php
@@ -2414,15 +2414,13 @@ class Imap extends ImapBodyStruct
 			$this->message_part_read+=strlen($line);
 		}
 
-		if ($this->message_part_size < $this->message_part_read) {
-
+		if ($line && $this->message_part_size < $this->message_part_read) {
 			$line = substr($line, 0, ($this->message_part_read-$this->message_part_size)*-1);
 		}
 
 		if($line===false) {
 			if($this->readFullLiteral) {
 				//don't attempt to read response after already have done that because it will hang for a long time
-				$this->readFullLiteral = true;
 				return false;
 			}
 
@@ -2448,7 +2446,7 @@ class Imap extends ImapBodyStruct
 	 * @param false $peek
 	 * @return bool
 	 */
-	public function save_to_file($uid, $path, $imap_part_id=-1, $encoding='', $peek=false) :bool
+	public function save_to_file($uid, $path, $imap_part_id=-1, $encoding='', $peek=true) :bool
 	{
 
 		$fp = fopen($path, 'w+');
@@ -2562,8 +2560,6 @@ class Imap extends ImapBodyStruct
 	public function copy(array $uids, $mailbox='INBOX') :bool
 	{
 		$this->clean($mailbox, 'mailbox');
-
-		$uid_string = implode(',',$uids);
 
 		$command = "UID COPY %s \"".$this->addslashes($this->utf7_encode($mailbox))."\"\r\n";
 		$status = $this->_runInChunks($command, $uids);
@@ -2858,7 +2854,7 @@ class Imap extends ImapBodyStruct
 
 		$this->clean($mailbox, 'mailbox');
 		$this->clean($size, 'uid');
-		$command = 'APPEND "'.$this->utf7_encode($mailbox).'" ('.$flags.') {'.$size."}\r\n";
+		$command = 'APPEND "'.$this->addslashes($this->utf7_encode($mailbox)).'" ('.$flags.') {'.$size."}\r\n";
 		$this->send_command($command);
 		$result = fgets($this->handle);
 		if (substr($result, 0, 1) == '+') {