+++ www/go/core/install/updates.php
@@ -1367,3 +1367,11 @@ $updates['202302211524'][] = "ALTER TABLE `core_pdf_template`
 
 $updates['202302211524'][] = "ALTER TABLE `core_pdf_template`
   MODIFY `id` bigint unsigned auto_increment";
+
+$updates['202303131003'][] = "delete from core_acl_group where groupId = (select id from core_group where isUserGroupFor=1);";
+
+
+$updates['202303151524'][] = "ALTER TABLE `core_user` 
+ADD COLUMN `themeColorScheme` ENUM('light', 'dark', 'system') NOT NULL DEFAULT 'light' AFTER `theme`;";
+
+$updates['202303151524'][] = "UPDATE `core_user` SET theme = 'Paper', themeColorScheme = 'dark' WHERE theme = 'Dark';";