+++ www/go/modules/community/googleauthenticator/language/he.php
@@ -1,0 +1,4 @@
+<?php
+return array (
+  'Block Group-Office usage until setup is done' => 'Block {product_name} usage until setup is done',
+);