+++ www/language/languages.php
@@ -20,6 +20,7 @@ $languages['es']='Español';
 $languages['et']='Estonian';
 $languages['el']='Ελληνικά';
 $languages['fr']='Francais';
+$languages['he']= '(Hebrew) עִבְרִית';
 $languages['hr']='Hrvatski';
 $languages['it']='Italiano';
 $languages['id']='Bahasa Indonesia';