+++ www/modules/email/MessagesGrid.js
@@ -543,7 +543,7 @@ Ext.extend(GO.email.MessagesGrid, go.grid.GridPanel,{
 		}
 
 		return unseen + icons.map(function(i) {
-			return '<i class="icon em-'+i+'">' + i + '</i>';
+			return '<i class="icon c-secondary">' + i + '</i>';
 		}).join("");
 		
 	},