+++ www/modules/email/language/he.php
@@ -1,0 +1,7 @@
+<?php
+return array (
+  'Search through Group-Office' => 'Search through {product_name}',
+  'Label' => 'ملصق',
+  'Ok' => 'موافقة',
+  'Contact' => 'جهة اتصال',
+);